fix: handle the closing of the modal in the modal itself
The backdrop can handle the escape key, but cannot properly handle the
click on the backdrop. What is required is a decision if the click
happened "outside" of the content, however the modal uses a bullseye
layout, which captures the click first.

The modal now handles escape and click and can evaluate if the click
happened inside or outside the modal.