refactor: move towards hook based popperjs and new menus
This change works towards replacing the popperjs component with a new,
hook/functional component based implementation. It also makes the
popper component for flexible, but also more complex. I am not happy
yet.

The main reason for this is the new menu/menu toggle system which
replaces most of the previously existing drop down style components
(like select, app-launcher, context-switcher).

Aside from the app-launcher, the other components still exists, but
are deprecated. The app-launcher is replaced with a new drop-down-menu.

This isn't finishes yet, but some work in progress. Still, it should
work and with a single exception be bug free. Having a lot of
limitations though.