fix: expansion of rows does not work
Right now, the callback of the table toggle model is cached. Which is
a good thing, as re-calling the use_table_data hook does not create
a new callback instance and thus does not require re-rendering the
table.

Unfortunately, there is no alternative trigger to re-render the changed
row either. So, expanding rows doesn't work.

This is a workaround until we properly handle expansion state per-row.