fix: Dispatch the autofocus effect on value of autofocus prop
Previously, the use_effect_with call was invoked with the NodeRef on
the respective components' Properties, which works well if user code
passes a stable NodeRef in as a prop; if the prop gets defaulted
though, that value changes every time through the re-render... which
would auto-focus the field everytime an unrelated field in the form
changed.

So instead, pass the value of the autofocus prop as an effect
condition, and move the input ref the traditional way (via
cloning/move closure capture). This ensures that the effect is
triggered only once on first render, and then only when the autofocus
prop changes.

Refs: #145
Fixes: #145