Promote the gallery elements' `key` attributes to their container
Unless every direct child of the gallery has a key element, every time
the set of children changes, *every* child will be re-rendered, which
leads to scroll state confusion and bad performance of the gallery.

So, let's promote up the child's `key` and if there's none set, set
one. I'm not super happy with that latter behavior and will try to
refine it further, so that the defaulted keys don't depend on the
order of the list.