Use the virtual dom to promote the key attribute upwards
This adds a util function `wrapper_div_with_attributes` that
constructs a div with as-static-as-we-can-make-them
attributes (similar to what html! does), but promoting up the child
element's `key` field, if one exists.

This kind of util function should be possible to be used in other
list-like elements that require wrapping of children (grid, for
example).